refactor: remove unused imports and simplify return statements in evaluation methods

+++ src/main/java/dicechess/bot/OnnxEvaluator.java
@@ -69,7 +69,7 @@ public class OnnxEvaluator implements AutoCloseable {
     public float evaluate(GameState state, int color) {
         if (!isLoaded || session == null || env == null) {
             // Engine heuristic fallback
-            return (float) Evaluator.evaluate(state, color);
+            return Evaluator.evaluate(state, color);
         }
 
         try {
@@ -97,7 +97,7 @@ public class OnnxEvaluator implements AutoCloseable {
             }
         } catch (OrtException e) {
             logger.log(Level.WARNING, "Error running ONNX inference: {0}. Falling back to engine evaluation.", e.getMessage());
-            return (float) Evaluator.evaluate(state, color);
+            return Evaluator.evaluate(state, color);
         }
     }
